+++ src/main/java/org/opentripplanner/graph_builder/DataImportIssuesToHTML.java
@@ -268,16 +268,18 @@ public class DataImportIssuesToHTML implements GraphBuilderModule {
                         label = label_name + currentCount;
                         if (label.equals(issueTypeName)) {
                             out.printf(
-                                    "<button class='pure-button pure-button-disabled button-%s'>%s</button>%n",
+                                    "<button class='pure-button pure-button-disabled button-%s' style='background-color: %s;'>%s</button>%n",
                                     label_name.toLowerCase(),
+                                    IssueColors.rgb(label_name),
                                     label
                             );
                         }
                         else {
                             out.printf(
-                                    "<a class='pure-button button-%s' href=\"%s.html\">%s</a>%n",
+                                    "<a class='pure-button button-%s' href=\"%s.html\" style='background-color: %s;'>%s</a>%n",
                                     label_name.toLowerCase(),
                                     label,
+                                    IssueColors.rgb(label_name),
                                     label
                             );
                         }