+++ src/main/java/org/opentripplanner/routing/algorithm/filterchain/GroupBySimilarity.java
@@ -6,18 +6,18 @@ import org.opentripplanner.model.base.ToStringBuilder;
  * Group itineraries by similarity and reduce the number of itineraries down to an given
  * maximum number of itineraries per group.
  * <p>
- * Group itineraries by the main legs and keeping approximately the given total number of
- * itineraries. The itineraries are grouped by the legs that account for more then 'p' % for the
+ * Group itineraries by the main legs and keeping at most the given total number of itineraries.
+ * The itineraries are grouped by the legs that account for more then 'p' % for the
  * total distance.
  * <p/>
  * If the time-table-view is enabled, the result may contain similar itineraries where only the
  * first and/or last legs are different. This can happen by walking to/from another stop,
  * saving time, but getting a higher generalized-cost; Or, by taking a short ride.
  * Use {@code groupByP} in the range {@code 0.80-0.90} and {@code approximateMinLimit=1} will
- * remove these itineraries an keep only the itineraries with the lowest generalized-cost.
+ * remove these itineraries and keep only the itineraries with the lowest generalized-cost.
  * <p>
  * When this filter is enabled, itineraries are grouped by the "main" transit legs. Short legs
- * are skipped. Than for each group of itineraries the itinerary with the lowest cost is kept.
+ * are skipped. Then for each group of itineraries the itinerary with the lowest cost is kept.
  * All other itineraries are dropped.
  * <p>
  * A good way to allow for some variation is to include several entries, relaxing the min-limit,
@@ -31,7 +31,7 @@ import org.opentripplanner.model.base.ToStringBuilder;
  * Normally, we want some variation, so a good value to use for this parameter is the combined
  * cost of board- and alight-cost including indirect cost from board- and alight-slack.
  *
- * @see ItineraryFilterChainBuilder#addGroupBySimilarity(GroupBySimilarity)
+ * @see ItineraryListFilterChainBuilder#addGroupBySimilarity(GroupBySimilarity)
  */
 public class GroupBySimilarity {
 
@@ -47,10 +47,55 @@ public class GroupBySimilarity {
    */
   public final int maxNumOfItinerariesPerGroup;
 
+  /**
+   * Should the grouped itineraries be further grouped based on the boarding and alighting stations
+   * or stops of all the legs in the itinerary, so that only one itinerary that has the same
+   * combination of stations/stops is shown.
+   */
+  public final boolean nestedGroupingByAllSameStations;
 
-  public GroupBySimilarity(double groupByP, int maxNumOfItinerariesPerGroup) {
+  /**
+   * Remove all itineraries whose cost for the non-grouped legs is at least this much higher
+   * compared to the lowest cost in the group.
+   */
+  public final double maxCostOtherLegsFactor;
+
+  /**
+   * Get a GroupBySimilarity configured to return one itinerary per group.
+   */
+  public static GroupBySimilarity createWithOneItineraryPerGroup(double groupByP) {
+    return new GroupBySimilarity(groupByP, 1, false, 0.0);
+  }
+
+  /**
+   * Get a GroupBySimilarity configured to return more than one itinerary per group.
+   * Additional filtering is configured with nestedGroupingByAllSameStations and maxCostOtherLegsFactor
+   * parameters,
+   */
+  public static GroupBySimilarity createWithMoreThanOneItineraryPerGroup(
+          double groupByP,
+          int maxNumOfItinerariesPerGroup,
+          boolean nestedGroupingByAllSameStations,
+          double maxCostOtherLegsFactor
+  ) {
+    return new GroupBySimilarity(
+        groupByP,
+        maxNumOfItinerariesPerGroup,
+        nestedGroupingByAllSameStations,
+        maxCostOtherLegsFactor
+    );
+  }
+
+  private GroupBySimilarity(
+          double groupByP,
+          int maxNumOfItinerariesPerGroup,
+          boolean nestedGroupingByAllSameStations,
+          double maxCostOtherLegsFactor
+  ) {
     this.groupByP = groupByP;
     this.maxNumOfItinerariesPerGroup = maxNumOfItinerariesPerGroup;
+    this.nestedGroupingByAllSameStations = nestedGroupingByAllSameStations;
+    this.maxCostOtherLegsFactor = maxCostOtherLegsFactor;
   }
 
   @Override
@@ -58,6 +103,8 @@ public class GroupBySimilarity {
     return ToStringBuilder.of(GroupBySimilarity.class)
         .addNum("groupByP", groupByP)
         .addNum("maxNumOfItinerariesPerGroup", maxNumOfItinerariesPerGroup)
+        .addBoolIfTrue("nestedGroupingByAllSameStations", nestedGroupingByAllSameStations)
+        .addNum("maxCostOtherLegsFactor", maxCostOtherLegsFactor)
         .toString();
   }
 }