+++ src/main/java/org/opentripplanner/transit/raptor/api/transit/RaptorTransitDataProvider.java
@@ -2,6 +2,7 @@ package org.opentripplanner.transit.raptor.api.transit;
 
 
 import java.util.Iterator;
+import javax.validation.constraints.NotNull;
 
 
 /**
@@ -51,13 +52,21 @@ public interface RaptorTransitDataProvider<T extends RaptorTripSchedule> {
      * </pre>
      * @return a map of distances from the given input stop to all other stops.
      */
-    Iterator<? extends RaptorTransfer> getTransfers(int fromStop);
+    Iterator<? extends RaptorTransfer> getTransfersFromStop(int fromStop);
+
+    /**
+     * This method is responsible for providing all transfers to a given stop from all
+     * possible stops around that stop.
+     * See {@link #getTransfersFromStop(int)} for detail on how to implement this.
+     * @return a map of distances to the given input stop from all other stops.
+     */
+    Iterator<? extends RaptorTransfer> getTransfersToStop(int toStop);
 
     /**
      * Return a set of all patterns visiting the given set of stops.
      * <p/>
      * The implementation may implement a lightweight {@link RaptorTripPattern} representation.
-     * See {@link #getTransfers(int)} for detail on how to implement this.
+     * See {@link #getTransfersFromStop(int)} for detail on how to implement this.
      *
      * @param stops set of stops for find all patterns for.
      */
@@ -73,4 +82,26 @@ public interface RaptorTransitDataProvider<T extends RaptorTripSchedule> {
      * Create/provide the cost criteria calculator.
      */
     CostCalculator multiCriteriaCostCalculator();
+
+
+    /**
+     * Implement this method to provide a service to search for {@link RaptorTransferConstraint}.
+     * This is not used during the routing, but after a path is found to attach constraint
+     * information to the path.
+     * <p>
+     * The search should have good performance, but it is not a critical part of the overall
+     * performance.
+     */
+    RaptorPathConstrainedTransferSearch<T> transferConstraintsSearch();
+
+    /**
+     * Raptor relies on stop indexes for all references to stops for performance reasons, but
+     * when a critical error occurs, it is nice to be able to inject information to the
+     * log event or during debugging to see which stop it is. This is important to be able to
+     * reproduce the error. This method is used by Raptor to translate from the stop index to a
+     * string which should be short and identify the stop given the related pattern, for example
+     * the stop name would be great.
+     */
+    @NotNull
+    RaptorStopNameResolver stopNameResolver();
 }