+++ src/test/java/org/opentripplanner/transit/raptor/_data/transit/TestTransitData.java
@@ -1,44 +1,53 @@
 package org.opentripplanner.transit.raptor._data.transit;
 
-import static org.opentripplanner.model.transfer.TransferConstraint.MAX_WAIT_TIME_NOT_SET;
-
 import java.util.ArrayList;
 import java.util.HashSet;
 import java.util.Iterator;
 import java.util.List;
 import java.util.Set;
+import java.util.stream.Collectors;
+import javax.annotation.Nullable;
 import lombok.val;
 import org.opentripplanner.model.transfer.ConstrainedTransfer;
 import org.opentripplanner.model.transfer.TransferConstraint;
-import org.opentripplanner.model.transfer.TransferPriority;
 import org.opentripplanner.routing.algorithm.raptor.transit.cost.DefaultCostCalculator;
 import org.opentripplanner.routing.algorithm.raptor.transit.cost.McCostParamsBuilder;
 import org.opentripplanner.routing.algorithm.transferoptimization.model.TripStopTime;
 import org.opentripplanner.routing.algorithm.transferoptimization.services.TransferServiceAdaptor;
+import org.opentripplanner.transit.raptor._data.RaptorTestConstants;
 import org.opentripplanner.transit.raptor._data.debug.TestDebugLogger;
 import org.opentripplanner.transit.raptor.api.request.RaptorRequestBuilder;
 import org.opentripplanner.transit.raptor.api.transit.CostCalculator;
 import org.opentripplanner.transit.raptor.api.transit.IntIterator;
+import org.opentripplanner.transit.raptor.api.transit.RaptorConstrainedTransfer;
+import org.opentripplanner.transit.raptor.api.transit.RaptorPathConstrainedTransferSearch;
 import org.opentripplanner.transit.raptor.api.transit.RaptorRoute;
+import org.opentripplanner.transit.raptor.api.transit.RaptorStopNameResolver;
 import org.opentripplanner.transit.raptor.api.transit.RaptorTransfer;
 import org.opentripplanner.transit.raptor.api.transit.RaptorTransitDataProvider;
+import org.opentripplanner.transit.raptor.util.ReversedRaptorTransfer;
 
 @SuppressWarnings("UnusedReturnValue")
-public class TestTransitData implements RaptorTransitDataProvider<TestTripSchedule> {
+public class TestTransitData implements RaptorTransitDataProvider<TestTripSchedule>, RaptorTestConstants {
 
-  private static final TransferConstraint GUARANTEED = new TransferConstraint(
-          TransferPriority.ALLOWED, false, true, MAX_WAIT_TIME_NOT_SET
-  );
+  private static final TransferConstraint GUARANTEED = TransferConstraint.create()
+          .guaranteed().build();
 
-  private final List<List<RaptorTransfer>> transfersByStop = new ArrayList<>();
+  private final List<List<RaptorTransfer>> transfersFromStop = new ArrayList<>();
+  private final List<List<RaptorTransfer>> transfersToStop = new ArrayList<>();
   private final List<Set<TestRoute>> routesByStop = new ArrayList<>();
   private final List<TestRoute> routes = new ArrayList<>();
   private final List<ConstrainedTransfer> guaranteedTransfers = new ArrayList<>();
   private final McCostParamsBuilder costParamsBuilder = new McCostParamsBuilder();
 
   @Override
-  public Iterator<? extends RaptorTransfer> getTransfers(int fromStop) {
-    return transfersByStop.get(fromStop).iterator();
+  public Iterator<? extends RaptorTransfer> getTransfersFromStop(int fromStop) {
+    return transfersFromStop.get(fromStop).iterator();
+  }
+
+  @Override
+  public Iterator<? extends RaptorTransfer> getTransfersToStop(int toStop) {
+    return transfersToStop.get(toStop).iterator();
   }
 
   @Override
@@ -64,6 +73,38 @@ public class TestTransitData implements RaptorTransitDataProvider<TestTripSchedu
     );
   }
 
+  @Override
+  public RaptorPathConstrainedTransferSearch<TestTripSchedule> transferConstraintsSearch() {
+    return new RaptorPathConstrainedTransferSearch<>() {
+      @Nullable
+      @Override
+      public RaptorConstrainedTransfer findConstrainedTransfer(
+              TestTripSchedule fromTrip,
+              int fromStopPosition,
+              TestTripSchedule toTrip,
+              int toStopPosition
+      ) {
+        var list = routes.stream()
+                .flatMap(r -> r.listTransferConstraintsForwardSearch().stream())
+                .filter(tx -> tx.getSourceTrip().equals(fromTrip))
+                .filter(tx -> tx.getSourceStopPos() == fromStopPosition)
+                .filter(tx -> tx.getTrip().equals(toTrip))
+                .filter(tx -> tx.getStopPositionInPattern() == toStopPosition)
+                .collect(Collectors.toList());
+
+        if(list.isEmpty()) { return null; }
+        if(list.size() == 1) { return list.get(0); }
+        throw new IllegalStateException("More than on transfers found: " + list);
+      }
+    };
+  }
+
+  @Override
+  public RaptorStopNameResolver stopNameResolver() {
+    // Index is translated: 1->'A', 2->'B', 3->'C' ...
+    return this::stopIndexToName;
+  }
+
   public TestRoute getRoute(int index) {
     return routes.get(index);
   }
@@ -75,6 +116,7 @@ public class TestTransitData implements RaptorTransitDataProvider<TestTripSchedu
       debug.addStops(stopsVisited());
     }
     val logger = new TestDebugLogger(true);
+
     debug
         .stopArrivalListener(logger::stopArrivalLister)
         .patternRideDebugListener(logger::patternRideLister)
@@ -101,8 +143,9 @@ public class TestTransitData implements RaptorTransitDataProvider<TestTripSchedu
   }
 
   public TestTransitData withTransfer(int fromStop, TestTransfer transfer) {
-    expandNumOfStops(fromStop);
-    transfersByStop.get(fromStop).add(transfer);
+    expandNumOfStops(Math.max(fromStop, transfer.stop()));
+    transfersFromStop.get(fromStop).add(transfer);
+    transfersToStop.get(transfer.stop()).add(new ReversedRaptorTransfer(fromStop, transfer));
     return this;
   }
 
@@ -126,6 +169,7 @@ public class TestTransitData implements RaptorTransitDataProvider<TestTripSchedu
     }
     guaranteedTransfers.add(
         new ConstrainedTransfer(
+            null,
             new TestTransferPoint(fromStop, fromTrip),
             new TestTransferPoint(toStop, toTrip),
             GUARANTEED
@@ -163,7 +207,7 @@ public class TestTransitData implements RaptorTransitDataProvider<TestTripSchedu
         return findGuaranteedTransfer(from.trip(), from.stop(), toTrip, toStop);
       }
     };
-  };
+  }
 
 
   /* private methods */
@@ -175,7 +219,8 @@ public class TestTransitData implements RaptorTransitDataProvider<TestTripSchedu
 
   private void expandNumOfStops(int stopIndex) {
     for (int i=numberOfStops(); i<=stopIndex; ++i) {
-      transfersByStop.add(new ArrayList<>());
+      transfersFromStop.add(new ArrayList<>());
+      transfersToStop.add(new ArrayList<>());
       routesByStop.add(new HashSet<>());
     }
   }
@@ -189,5 +234,4 @@ public class TestTransitData implements RaptorTransitDataProvider<TestTripSchedu
     }
     return stops;
   }
-
 }